flask-ollama-proxy/requirements.txt
dicttoxml==1.7.4
Flask==3.1.0
Flask_Cors==4.0.1
flask_sqlalchemy==3.1.1
openai==1.61.1
pdfkit==1.0.0
pypandoc==1.15
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
Werkzeug==3.1.3
xmltodict==0.14.2
psycopg2-binary
Flask-Migrate
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
bcrypt==3.2.0
pyjwt<2.10
SQLAlchemy~=2.0.38
alembic~=1.14.1
imgkit
Pillow